pyproject.toml
[project]
name = "stirlingpdfassistant"
version = "0.1.0"
description = "A Telegram Bot Assistant for Stirling PDF"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.7",
]

[project.scripts]
stirling-bot = "stirlingpdf_assistant.main:main"

[dependency-groups]
dev = [
    "coverage>=7.13.5",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_default_fixture_loop_scope = "function"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/stirlingpdf_assistant"]
